feat: rewrite adapter and move CFW adapters to web
